[#1336] Show non-blank line count in the brackets instead of blank lines (#1347)
The report in the code panel shows the number of blank lines written 
by an author in a file or in all files of a file type inside the 
brackets beside the total number of lines written. The viewer needs 
to do mental calculation to get the number of non-blank lines 
written by the author.

Let's show non-blank line count instead, and update relevant user 
guide.