[#1345] Authorship: Add last modified date to each LoC if specified (#1348)
The last modified date of each line of code is not present in the
`authorship.json`.

It would be beneficial for tutors to use the last modified date of
each line of code to do further processing.

Let's add the last modified date of each line of code within the
`authorship.json`. This functionality will only be enabled when the
user specifies the flag '--last-modified-date' or '-l' in short.