[#436] Give warning for duplicate alias (#1187)
Duplicate aliases in the csv files or standalone config do not trigger 
any warning. As a result,the user might not get to know that there are 
duplicate aliases in the csv files provided by him/her.

Let's give a warning if there are duplicate aliases in the csv file or 
standalone config to notify the user that the alias will belong to the 
last person who claims it.