[#1207] AuthorConfiguration: Rename authorEmailsAndAliasesMap to authorDetailsToAuthorMap (#1362)
`authorEmailsAndAliasesMap` in the `AuthorConfiguration` has a
structure that does not match its name. It stores mappings to 
`author` from `git id`, `email`, and `alias` while the naming entails 
the mapping is from `email`s to `aliase`s.

Let's rename `authorEmailsAndAliasesMap` to 
`authorDetailsToAuthorMap` to improve clarity and better match the 
variable name to its functionality.