[#1202] Add Javadoc for backend functions (#1364)
Some methods in `StreamGobbler`, `CustomLogFormatter`, `FileType`, 
and `RepoConfiguration` do not have Javadoc although the methods are 
not standard getter or setter.

To enhance readability, it may be necessary to explain what they do, 
especially for `run()` in `StreamGobbler` since the method name may 
not give enough information. Additionally, there is a colon missing 
in the developer guide architecture section.

Let's add Javadoc to these methods and also add the missing colon
to the developer guide to make documentation more consistent.