Update documentation to remove need for personal access token
Currently, users need to specify a personal access token or a
deploy key to automate the generation of the RepoSense report.
This is an additional step that makes using this repository less
user-friendly for new users.

Let's update the documentation to remove the need for the user to
generate a personal access token, since the GITHUB_TOKEN is
automatically used when generating a report in GitHub Actions.