Clone test repositories only once when testing
The test repositories are cloned for every test that performs
cloning, which results in tests taking a long time to complete as
a significant amount of time is spent cloning the same repository
for every test.

Let's prevent the deletion of the cloned test repositories until
we are sure that all tests have completed, thus allowing the test
repositories to be cloned only once during each test run.