[#674] v_authorship: include binary files touched (#1413)
The authorship panel does not show the binary files touched by the 
author in the repository. This is because content changes in binary 
files are usually unreadable by the user, as is shown in #711. 

However, with the exclusion of binary files, the authorship panel 
may not accurately reflect the authorship contribution. It will be 
appropriate to list the title of the binary files without showing 
the content change to the user. This can make the authorship 
contribution summary more accurate and organized.

Let's include the binary files touched in the authorship panel.