Support defining sinceDate, untilDate and period in report config
The new report-config.json allows for the configuration of the
report generation. However, only the report title can be
configured at this time.

Let's expand the functionality of the report config file by
allowing sinceDate, untilDate and period to be defined as well.
If the CLI arguments were to contain the same settings as those in
the report config, then the CLI arguments will take precedence.