Implement shallow cloning (#1449)
RepoSense downloads the entire commit history of the repo when 
cloning it. 

This results in a lot of unnecessary history being downloaded, which 
increases the size of the cloned repo and slows down the time taken 
to clone the repo.

Let's implement Git's shallow cloning functionality so as to 
truncate the downloaded history thereby reducing the size of the 
cloned repo and speeding up the cloning process.