[#1485] Watch tab info directly (#1490)
Zoom tab does not update properly if one of the zoom watchables do 
not change.

Let's watch the `tabZoomInfo` stored in `Vuex` instead of 
`zoomOwnerWatchable`. This works as we ensure that every change to 
`tabZoomInfo` stored in `Vuex` replaces `tabZoomInfo` with a new 
object. 

Everything here also applies to the authorship tab with
`authorshipOwnerWatchable` and `tabAuthorshipInfo` instead.