CSVParser: check column headers when parsing config CSV files (#1459)
RepoSense parses the config CSV files purely based on column order. 

This means that every column needs to be present in the correct 
order (even if the column is optional) in order for the CSV
file to be parsed. This additionally means that every time we add a
new column to a config CSV, we would break compatibility with 
existing config CSV files for all users.

Let's change the `CSVParser` to verify and parse the config CSV 
files based on the column headers. This would allow us to parse 
config CSVs with columns in a different order or with some 
(optional) columns missing.