Improve assertion of authors for file analysis
The assertion used in checking the correctness of the attribution
of authors is not entirely robust and can be prone to other errors
which may not be detected.

Let's refactor the assertion used in file and annotation analysis
such that they use the same function and make the tests more
robust.