Allow disowning of code by using empty author tags
Empty author tags are assumed to represent the end of an
annotation block. However, there may be cases where certain code
blocks should not be attributed to any author. The current
workaround is to attribute the code to some random author, which
is not ideal from the user's perspective.

Let's implement a new functionality for users to disown code by
using empty author tags.