Move logic for getting ReportConfiguration to ArgsParser
The current design of getting the ReportConfiguration object in
the main RepoSense class is flawed and the design is not ideal,
as it meant that configuration has to be set and parsed twice.

Let's move the logic for getting the ReportConfiguration object
into ArgsParser to avoid having to duplicate configuration parsing
in the future.