[#1486] Frontend: rename zIsMerge to zIsMerged (#1487)
The zoom panel of a merged group currently cannot be restored after 
reloading the page. This is caused by incomplete renaming of the 
zIsMerge argument in #1457.

Renaming other instances of zIsMerge can resolve this issue and make 
the reloading behavior more consistent with the user expectation.

Let's change all instances of zIsMerge to zIsMerged.