Add link to path glob syntax in the User Guide (#1541)
The path glob syntax used in the Ignore Glob List column of the CSV 
configuration files is not documented for the user. This can cause 
issues when an invalid path glob is used to generate the report.

Adding a link to the official documentation will make it more 
convenient to the user.

Let's add the link to the glob syntax in the User Guide.