Add warning when there are a large number of files to be analyzed (#1524)
If RepoSense is run on a repo with a large number of text files 
(with a large number of lines overall), it can take a long time to 
generate the authorship report. Additionally, the size of the 
`authorship.json` file can be very large, possibly even exceeding 
the file size limit of 100 MB for GitHub, which would prevent the 
report from being deployed on GitHub Pages.

Let's add a warning when there are a large number of text files with 
a large number of lines overall. This would let users know the 
reason why they may experience slow performance and would also 
advise users on how they can reduce the number of files to be 
analyzed using `repo-config.csv`.