[#1435] Expand continuous integration testing to Ubuntu 18.04 and 20.04 (#1503)
The continuous integration test suite only runs on Ubuntu 16.04,
which is a relatively older version.

Let's expand the testing to both Ubuntu 18.04 and 20.04 to make
sure that RepoSense works for both environments as well. This is
done by defining a job matrix that can automatically run the same
build steps for all versions of Ubuntu.