[#1512] Seemingly random disappearance of dates (#1513)
There is a possibility that in `v_summary.js`, both `minDate`
and `tmpFilterSinceDate` are empty strings. This leads to bugs where 
both  `minDate` and `tmpFilterSinceDate` cannot be set.

Let's set `minDate` to `window.app.sinceDate`. This is possible as 
`api.loadSummary()` must be run in order for `userUpdated` to be set 
to true in 'main.js`, and hence `v-summary.js` will not load before 
this.

Let's remove the `getDates()` function in `v_summary`. This is 
possible as `tmpFilterSinceDate` does not need to be set, as its 
only dependency is `filterSinceDate` which checks that 
`tmpFilterSinceDate` is empty.