[#954] RepoConfiguration: remove redundant usage of streams (#1558)
In the setStandaloneConfigIgnoredToRepoConfigs method of
RepoConfiguration, we iterate through a list of RepoConfiguration
objects and modify the isStandaloneConfigIgnored attribute in each
object. Currently, we use Stream.forEach(), which can be replaced by a
list.forEach() method since we already have a list.

Let's replace the redundant Stream.forEach() with the .forEach()
method.