[#937] Make all tests throw generic Exception (#1566)
The current test code for RepoSense contains throws clauses that do
not throw a generic Exception. This can result in a series of code
changes that have to be made when the exception(s) thrown by an
existing method used in the test code is changed, to prevent a
compiler error.

Let's make all methods in the test (and systemtest) folder(s) of
RepoSense throw a generic Exception.