Cypress: replace nonexistent input element with pug id in selector in codeView_reload.js (#1580)
Cypress test in codeView.js is failing because the test analyzes
RepoSense's repository itself to run tests on. When PR#1412
was merged in, RepoSense no longer has any pug files whereas
the test case assumes pug files are still there thus leading to an
error. 

The selector utilized in the test refers to an input element with id 
`pug` that no longer exists.

Let's change the selector in the test to refer to an input element
with id `java` that still exists in RepoSense's repository.