Update DG report
After the transition to Vue, there are links and diagrams in
the DG HTML report page that refer to files that previously
had .js extension and/or were located in now-deleted folders.

Let's update these links and diagrams to reflect the new file
extensions and the corresponding location of these files.