[#1463] Fix case insensitive bug in @@author feature (#1577)
Fix case insensitive bug of Map object used in @@author feature 
implementation

In AuthorConfiguration.java, the authorDetailsToAuthorMap object
utilized in AnnotatorAnalyzer.java for implementing the @@author
feature marks githubIds and gitIds with the same character sequence
but different casing as duplicates resulting in unnecessary warnings
in the log. When duplicates are encountered previous mappings might
also be altered resulting in potential unexpected assignings when
using the @@author feature.

Let's change the type of Map associated with the
authorDetailsToAuthorMap object from a TreeMap with
String.CASE_INSENSITIVE_ORDER as its Comparator to a HashMap.