[#382, #1568] Make annotation tagged author appear in report (#1554)
The current implementation will mark tagged authors as unknown if
their names do not appear in the commit history of the repo or in the
author config file.

Let's show all author introduced via the @@author tags if there is no
author config file. If there is an author config file, @@author tags
should not result in the author appearing in the report unless the
author is listed in the author config file.