[#1571] Fix incorrect invalidation of file path with spaces (#1576)
The "git diff" command on files where the filename contains a space
causes the resulting output to generate an additional \t character at
the end of the filename. This becomes an invalid file path in Windows
while being incorrectly ignored in Macs and Linux.

Let's fix the incorrect identification of valid file paths with
spaces.