[#1563, #1579] Tests: allow a fresh clone for some test cases (#1570)
The test repositories are cloned only once during testing, and a full
clone is performed. This causes certain test cases that require a
shallow clone to be tested incorrectly on a full clone.

Let's allow individual tests to specify if they wish to obtain a
fresh clone of a repository.