[#1606] GitHub Actions: update list of runners (#1608)
The Ubuntu 16.04 LTS runner for GitHub Actions is no longer supported
as at 16 October 2021. This is a breaking change that caused all
GitHub Actions checks to not complete.

The macOS 11 runner is also available for use now, and will be made
default on 13 December 2021.

Let's update the list of runners based on these changes.