Put repoId in commit
Using user.repoId for getBaseLink does not work when
the repoId gets overwritten by another repoId for which
the commit is invalid in a group merge.

Let's put repoId in the commit itself. This should give a valid
slice.repoId since getLink() for v-ramp takes in a commit as slice.