[#1621] Docs: case 3 report and settings links do not match (#1623)
Some of the links in showcase.md do not match the links given in the
address bar. In some cases, the links do not match the respective
descriptions or repository. 

In addition, some reports were generated using an older version of
RepoSense, which results in different URL parameters.

Let's change these links and use newer reports for consistency in the
URL style.