[#1638] Replace parseInt with Math.floor for contribution bars (#1639)
To calculate the number of contribution bars to display in 
getContributionBars, a parseInt of the ratio of a user's contribution
to the contribution limit is used. However, if the contribution is
much lower than average, the ratio is represented in scientific
notation and parseInt does not return the integer part of the number
(parseInt of 3.2301723943006836e-7 would return 3).

Let's use Math.floor to get the integer part of the contribution ratio 
instead of parseInt in getContributionBars, as is currently done in 
getFileTypeContributionBars.