[#1603] Terminate if a CSV config file has an unknown column (#1609)
RepoSense does not show any error if a config file has an unknown
column as it is likely to be a user mistake that could affect the
output.

Let's add a new exception (InvalidHeaderException) to handle this, as
using InvalidCsvException will not terminate the program. The check is
done within CsvParser::validateHeader.

The original code already handles case insensitive matching of column
names.