[#1539] Fix console error when doing individual merge group (#1618)
Passing slice.repoId to getBaseLink() after merging groups causes the
console error to happen. This is due to
window.REPOS[slice.repoId].location being undefined.

Let's put repoId in the commit itself and rename the "slice" argument
to "commit". This should give a valid commit.repoId since getLink()
for v-ramp takes in a commit. This would also allow the usage of
commit.repoId for separate users instead of user.repoId.