[#1615] Clear deployments for closed pull requests (#1654)
Deployment previews generated for open PRs are not cleared
automatically when a PR is closed or merged. As a result, environments
need to be cleared manually periodically, which is less than ideal.

Let's add a GitHub workflow for closed (which includes merged) PRs
that deletes generated preview environments automatically, and posts 
a comment on the PR with the preview links for later reference.