[#1648] Frontend: display reportTitle as browser tab title (#1656)
The title of the browser window displays 'reposense' instead of the 
`reportTitle` as specified in report-config.json. This is becuase the 
default index page title in the backend is wrongly specified. 

Let's correctly specify the default index page title, so that it gets 
correctly replaced by the specified reportTitle and shown in the browser
tab.