Actions: use pull_request_target for delete deploy workflow (#1675)
The delete deployments workflow merged in #1654 doesn't work for PRs 
from forked repos due to insufficient GitHub token permissions.

Additionally, the pending and surge workflows use fetch-depth: 0 when
checking out the repository, which unnecessarily fetches the full
history of the repository and cause the workflow to be slightly slower.

Let's use the pull_request_target workflow (instead of pull_request) 
which provides necessary permissions for the delete task, and give the
task a name. Let's also remove the unnecessary fetch-depth: 0 from the
pending and surge workflows.