Frontend: add Object.assign() to hash assignment (#1678)
The current implementation will result in some race condition where some
data are read and written at the same time. This is an issue in Vue 3 as
it does not wait for current function to finish before running another
async function.

Let's avoid this race condition by using Object.assign().