[#1659] Separate backend and frontend responsibilites (#1691)
The backend modifies files generated by the frontend. This breaks the
separation of responsibilities for generated report files, which makes 
testing more challenging.

Let's stop the backend manually modifying report titles in frontend
generated files, and have the frontend update the report title itself. 
The default report title generated by the frontend is now also 
'RepoSense Report' instead of 'reposense' so that there is no delay in
updating the report title in the default case. The backend will proceed
to generate report files in the absence of `templateZip.zip`. Let's
also add a test to check that report titles are correctly updated.