Upgrade Gradle to 6.9.2 (#1664)
Gradle has been stuck at 5.2.1 for quite some time. This version is not
compatible with a future migration to Java 17. However, upgrading to
Gradle 7.4 cannot be done yet due to a need to resolve errors from
third-party plugins.

Let's upgrade Gradle to 6.9.2 in the meantime to close the gap between
the version we are using and the latest available version.