[#1682] Fix space in local repo causing cloning exception (#1685)
Output directories are not quoted which can lead to errors. The 
GitClone::cloneFromBareAndUpdateBranch method is also not standardized
to match other clone methods which always run from the working
directory. This can be difficult to maintain.

Let’s fix the issue of spaces in repo directory causing clone failure
and standardize the cloneFromBareAndUpdateBranch method.