[#1736] Temporary adjust cypress test codeView_switchAuthorship
Due to the removal of the appveyor.yml file in #1731, the
codeView_switchAuthorship cypress test no longer works, as it expects
yml files to be present in the RepoSense repository.

Let's temporarily adjust the codeView_switchAuthorship cypress test so
that the frontend CI tests can pass, by removing the check on yml files.
In the longer term, the frontend tests' dependency on the RepoSense
repository should be removed, as it can make the CI tests unstable.