[#1738] Remove dependence of Cypress tests on RepoSense master (#1741)
Cypress tests depend on RepoSense master which is constantly changing.
This can break tests as it did when #1731 was merged.

Let's use a separate branch 'cypress' in the RepoSense repository to
generate fixed reports for frontend tests that are not affected by
merges to master. Changes required for testing new frontend features
can be made to the cypress branch which will be independent of master.
Let's also remove the AppVeyor CI badge since we no longer use it as
of PR #1731.