[#1630] Rearrange Gradle tasks and their dependencies (#1631)
When running pure backend tests alone using gradlew test systemtest, the
frontend is built before these tests. This is due to zipReport task
being needed for compileJava. zipReport in turn relies on buildFrontend.

The frontend is not used for these tests. It is wasteful to run 
buildFrontend when only the backend tests are being run.

In addition, some other tasks and properties are scattered throughout
the build.gradle file or have dependencies that could be pushed forward.
This complicates maintenance.

Let's remove the `test` and `systemtest` dependency on the frontend as
well as re-order and consolidate some task properties and dependencies.
This should make running either of them faster.