Docs: update list of project team members and commitment periods
The list of project team members and their commitment periods are
outdated, and does not include the current batch of contributors and
committers from CS3281 and CS3282. In addition, some project team
members may have already left and are not longer part of the team.

Let's update the documentation to reflect the changes in the project
team.