Refactor Json comparison methods and resources in systemtest (#1745)
The current structure of the system tests is such that it is meant to
only support the ConfigSystemTest. As such, the Json comparison methods
in ConfigSystemTest are not reusable. The resources used are also not
properly grouped.

This change aims to refactor some of the more useful Json comparison
methods so that they can be reused for future system tests.
Additionally, the resources are now grouped by the class file that is 
using them. 

Let's refactor the methods and resources in systemTest to make it easier
to implement new system test cases.