Refactor GitClone::clone method to increase flexibility in usage (#1749)
There are various methods in GitClone that are used only in testing.
However, they are still part of the main code and frequently updated.

This change would refactor the clone to cloneAndBranch, and move these
methods to a testing utility class to reduce the maintenance burden. A
more generic GitClone::clone method is also added in its place for
flexible usage during testing.

Let's move the GitClone::clone and cloneBare method to a testing class
to improve the ease of maintenance and implement a more generic clone
method for more flexible usage.