Update Node dependencies and fix JS code
Many dependencies are not up to date with the latest versions. This
gives rise to vulnerabilities as well as dependency conflicts when
attempting to run npm update.

In addition, there are ESLint checks for Airbnb JS standards which
fail upon updating these dependencies.

Let's update the dependencies and refactor the Javascript code
accordingly.