Fix arbitrary commit date and its timezone
The use of Instant.ofEpochMilli(Long.MIN_VALUE) as the arbitrary first
commit date now fails shallow cloning for
testSinceBeginningDateRangeWithShallowCloning system test.

In addition, using this arbitrary date in convertToGitDateRangeArgs
when converting timezone results in an overflow causing sinceDate to
be greater than untilDate. Thus, no commit results are retrieved,
leading to failing system tests.

Finally, the arbitrary first commit date is in UTC. However, in the
convertToGitDateRangeArgs method, the zoneId used for the conversion
is from the config, which may be different. This leads to an erroneous
timezone conversion.

Let's change Instant.ofEpochMilli() to use 0 so that the arbitrary date
is 1970-01-01. At the same time, let's convert the arbitrary date to
the correct timezone within the ArgsParser class and add a component
test for parsing --since d1.

This ensures that an arbitrary since date is adjusted to the
correct timezone for each RepoConfiguration before cloning any repos.