Update coding standard link and description
The oss-generic link for the coding standard is obsolete. At the same
time, the description vaguely refers to a module.

Finally, the checkstyle file also references our own standards.

Let's change the coding standard link to the one used by SE-EDU,
remove the reference to the module and add the link to our style
guide.