[#452] RepoConfiguration: Restrict sinceDate and untilDate
The flags sinceDate and untilDate are not checked to ensure they are
between 1970/01/01 00:00:00UTC and 2099/12/31 23:59:59UTC (inclusive)

Git is unable to handle timestamps before the Unix epoch and after 2099.
This had caused GitLogTest unit tests to fail. Upon investigation, this
problem applies across git commands, and not only for git log.

Therfore, sinceDate and untilDate should always range between 1970UTC
and 2099UTC for git commands to function expectedly.

Let's create a new method in TimeUtils to get a validated time. If the
input time does not lie in the range, we'll log a warning and then reset
the date to the latest or earliest permitted time. This function will
be used when first parsing CLI arguments, and also in the get methods of
sinceDate and untilDate to ensure the validaity of the two dates
everywhere.