Fix broken authorship panel-heading link after reload
After reloading the report with authorship panel open,
the header link of the authorship panel becomes unresponsive.

This is because the location key under the tabAuthorshipInfo object
 (a global state) is never set after reloading,
which causes the link to point to undefined.

Let's add a location key under the method that is responsible
for reloading the authorship panel so that the link
 points to proper repository.