Maintain file type label after filtering
When a commit contains two or more file types and
one of them is filtered out, the commit could still
be visible if it has other file types that are still
filtered in.

However, in such cases, the filtered-out file type label
is no longer visible for that commit. This could confuse
the users by indicating that the commit no longer consists
of that filtered-out file type.

Let's retain all the file types of a commit by returning
all the file types if at least one of them is still
filtered in and hence viewable to the users.