[#1726] Use respective icons for known repository types (#1800)
Currently, 'database' is used as the generic icon for all repository 
types.

As we allow more repository types, it will be clearer to the user if 
the icon reflects the repository used. Currently we are supporting 
'GitHub', 'GitLab', and 'BitBucket'. 'database' icon is kept as the 
default icon for repositories that does not belong to any of the above 
mentioned.

Let's use the respective icons for known repository types.