Update Cypress testing on zoom view
When a certain file type is unchecked under the zoom
view, the 'all' option is automatically unchecked,
and it assumes that unchecked file type will not appear
in the file labels under any commit.

However, a fixed behavior is that the unchecked file label
could still appear if that commit consists of other file types
that are still part of the checked file types.

Let's update Cypress testing on this behavior by checking
only 'all' option is unchecked.