Checkstyle: add MissingJavadocType check and update comments (#1781)
The oss-generic link for the coding standard is obsolete. At the same
time, the description vaguely refers to a module. The checkstyle file 
also references our own standards, which should be mentioned explicitly.

The JavadocType check actually checks that the Javadoc for classes,
enums and interfaces are of the correct format. The MissingJavadocType
check should be used instead to check that the public classes, enums and
interfaces have Javadoc.

Finally, the phrase "format's Javadoc" for the comment on the
JavadocMethod check does not make sense.

Let's change the coding standard link to the one used by SE-EDU, remove
the reference to the module and add the link to our style guide.

Let's also update the comments for the checks and add the
MissingJavadocType check.