[#1674] Convert zoneId String to ZoneId Object (#1790)
zoneId is stored as ZoneId object in CliArguments.java. It is converted 
into String when stored in RepoConfiguration.java. Subsequently, it is 
then converted back to ZoneId when required in FileInfoAnalyzer.java or 
other places.

Such repeated conversion is redundant.

Let's standardize and save zoneId as ZoneId Object across all files.