[#1661] Upgrade to Gradle 7 (#1780)
The Liferay plugin has recently been updated with annotations for
Gradle 7. As Gradle 7 is now available for use, this also means that the
Shadow plugin can be updated to the latest version.

There are some deprecated features that will be removed in a future
Gradle 8 release. These include an implicit dependency between
processResources and zipReport and the change from Report.enabled to
Report.required.

The processSystemtestResources task fails because Gradle 7 requires a
duplicatesStrategy to be defined for the task, which is of the Copy
type.

Let's upgrade to Gradle 7, update the relevant plugins and fix the
duplicatesStrategy error as well as deprecations related to Gradle 8.