Improve visualization on different sort parameters
Under authorship panel, when the files are sorted
by different parameters, there is no visual indication
that files are sorted in that parameter.

Let's improve the indication by placing the sort parameter
in front of the title so that the users know the sorting
parameter they are using to view the files.