[#1726] Update author config (#1799)
Previously, Reposense is only used with GitHub, hence 
"Author's GitHub ID" was used as the header in author-config.csv.

This header should be updated as we expand Reposense to be compatible 
with other platforms. Meanwhile, we would also like to still support 
the older version of author-config.csv that uses "Author's GitHub ID".

Let's update the header of author-config.csv to "Author's Git Host ID" 
while maintaining backward compatibility to the older version that 
still uses the "Author's GitHub ID" header.