[#1538] Frontend: make glob input wider (#1803)
There is plenty of space left beside the glob filter input box. We can
increase the box size to make use of the remaining spaces.

The glob filter input box and the filter button also does not align
properly after the increase in the box width when the viewport is
smaller.

Let's increase the width of the glob filter input box and fix the
alignment of the filter button.