[#1754] Ignore case when comparing author usernames (#1814)
Currently, identifying the author of a commit is done by getAuthor()
method in AuthorConfiguration.java. This matching is case-sensitive,
while GitHub usernames are case insensitive.

The function will not return the correct author if the input username is
case-wise different to the actual username.

Let's ignore case when identifying the author of a commit based on
author's aliases.