Add a pop-up that explains the meaning of the percentage label
When the groups are sorted by the amount of contribution,
a percentage label appears; however, no other information is
provided for that number, which could be confusing to the viewers.

Let's add a pop-up beside the percentage label so that the viewers
are aware that the percentage actually represents percentile for
each of the groups.