[#1645] Use RepoSense::main in ConfigSystemTest (#1816)
The current setup with the ConfigSystemTest::generateReport method 
bypasses the parsing of the original CLI arguments and setting up of 
RepoConfigurations.

If the main method is changed, then this method would also have to be 
updated, which can be troublesome. It is more beneficial to use 
RepoSense::main instead for the full end-to-end validation.

Let's rewrite the runTest method to use RepoSense::main method.