[#1765] Change npm version requirement to Node.js (#1786)
Node.js is often installed together with npm. With regards to tracking
version LTS and EOL, Node.js is often used instead of npm. Thus, it is
easier to update the Node.js version in the DG whenever the existing
version goes EOL and we upgrade to a newer version.

On the other hand, when updating the corresponding npm version, there
is a need to look for the Node.js version and then find the
corresponding npm version.

Let's change the DG pre-requisites to reference the Node.js version and
specify that it is for npm. This improves maintainability of the DG.