[#1810] Fix filter glob user input (#1827)
The input glob previously does not allow user input. Once clicked, the
report will refresh and does not allow user to type in anything in the
input box.

Let's remove v-on:focus to stop listening for focus event and stop
triggering indicateSearchBar(), which causes an update to
selectedFileTypes when the user clicks on the input box, resulting in
the report refreshing.

Let's also update input#search to also trigger indicateSearchBar() on
change event. input#search should have the same behaviour when the user
presses the Enter key as when user clicks button#submit-button.