[#1689] Disable links when they are broken (#1830)
There are some links in the report to the report repos such as view
group repo. However, when the link is missing or not supported by
RepoSense (i.e. is not Github, GitLab or Bitbucket), it points to
a non-existent link.

Let's disable the link when it is broken and change the message when the
users hover over the link to "This remote link is unsupported", so that
the users will know that the link is broken.