[#1813] Fix git clone exception when cloning lfs files (#1826)
When cloning repositories that utilize git-lfs, an exception will be
raised when the user's Git client is not properly configured.

Let's fix the exception that occurs while generating reports for
repositories that are using git-lfs. This is done by setting the Git
configuration to skip retrieving lfs files, and reset to user's original
configuration after the clone is complete.