[#1499] Declare vue prop types explicitly (#1852)
In the current version of RepoSense, props are declared explicitly for
Vue components as an array of strings. 

Let's declare props using the object syntax so that we can document the
prop types.

This will also show a warning in the browser console if props of the
wrong type are passed to the component. 

Let's also define classes for `Segment` and `User` for stricter type 
validation.