Codecov: specify threshold to be 0.3% (#1859)
By default, the threshold for code coverage is 0%, which means that any
decrease in the code coverage will cause the status checks to fail. This
can be confusing to contributors and are often ignored as false
positives.

Let's set a threshold of 0.3% based on the recent commits made to the
repository, to reduce the rate of false positives.