[#1791] Switch to package-lock.json and npm ci (#1863)
Issue 1784 demonstrates how changes to upstream technologies and
dependencies can break CI.  Although it was more likely to be due to
Chrome updating to version 102, there is also a possibility that
updating a dependency during a CI run might cause a breaking change. 

Let's consider switching to using package-lock.json and npm ci to ensure
that dependency versions are fixed when running CI. This allows for CI
builds to be repeatable and consistent.