[#968] Enforce Vue option order (#1867)
Currently, the order of component options is not enforced, resulting in
inconsistent ordering across files. Although there is a rule in the
ESLint Vue plugin that enforces this, it is currently being omitted.

Let's enable the order-in-components rule to enforce a consistent order
of component options according to the recommended order in the Vue style
guide.

This will improve the consistency and readability of the code.