Refactor integration.yml
The current implementation of integration.yml is difficult to update
since steps and version numbers are duplicated across operating systems
and the cypress workflow.

At the same time, between Windows and Mac/Ubuntu workflows, the $HOME
directory is indicated differently.

Let's consolidate the OS workflows and use environment variables to
keep track of the OS and tool versions across the steps as well as
standardise the $HOME directory nomenclature. This makes changing the
versions easier and also improves readability.