[#1880] Change hasAuthorConfigFile to become a local variable (#1910)
AuthorConfiguration.hasAuthorConfigFile is currently a static variable 
that determines whether during the analysis of repo files, a new author
should be added and tracked. If an author config file is provided, an
unrecognized author is ignored, otherwise it is added and tracked.

This might cause parallel runs of test cases at the method level, 
especially system test cases, to have issues.

Let's make AuthorConfiguration.hasAuthorConfigFile a local variable to
the AuthorConfiguration found within each RepoConfiguration.